Autorizacion Proyectos

+++ app/Http/Controllers/API/ProyectoController.php
@@ -6,10 +6,20 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\ProyectoResource;
 use App\Models\Proyecto;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 use PhpParser\Node\Stmt\TryCatch;
 
 class ProyectoController extends Controller
 {
+
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show']),
+        ];
+    }
+
     public $modelclass = Proyecto::class;
     /**
      * Display a listing of the resource.
@@ -31,8 +41,11 @@ class ProyectoController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Proyecto::class);
         $proyecto = json_decode($request->getContent(), true);
-
+        if (!$request->user()->esAdmin()) {
+            $proyecto['docente_id'] = $request->user()->id;
+        }
         $proyecto = Proyecto::create($proyecto);
 
         return new ProyectoResource($proyecto);
@@ -51,6 +64,8 @@ class ProyectoController extends Controller
      */
     public function update(Request $request, Proyecto $proyecto)
     {
+        Gate::authorize('update', Proyecto::class);
+
         $proyectoData = json_decode($request->getContent(), true);
         $proyecto->update($proyectoData);
 
@@ -62,6 +77,8 @@ class ProyectoController extends Controller
      */
     public function destroy(Proyecto $proyecto)
     {
+        Gate::authorize('delete', Proyecto::class);
+
         try {
             $proyecto->delete();
             return response()->json(null, 204);
